Move AtomicCounter inside _delete_all_values
