+++ sdk/python/feast/infra/online_stores/datastore.py
@@ -16,7 +16,7 @@ import logging
 from datetime import datetime
 from multiprocessing.pool import ThreadPool
 from queue import Queue
-from threading import Thread
+from threading import Lock, Thread
 from typing import Any, Callable, Dict, Iterator, List, Optional, Sequence, Tuple
 
 from pydantic import PositiveInt, StrictStr
@@ -35,7 +35,6 @@ from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
 from feast.protos.feast.types.Value_pb2 import Value as ValueProto
 from feast.repo_config import FeastConfigBaseModel, RepoConfig
 from feast.usage import log_exceptions_and_usage, tracing_span
-from feast.utils.generic_utils import AtomicCounter
 
 LOGGER = logging.getLogger(__name__)
 
@@ -272,6 +271,17 @@ def _delete_all_values(client, key):
     Creates and uses a queue of lists of entity keys, which are batch deleted
     by multiple threads.
     """
+
+    class AtomicCounter(object):
+        # for tracking how many deletions have already occurred; not used outside this method
+        def __init__(self):
+            self.value = 0
+            self.lock = Lock()
+
+        def increment(self):
+            with self.lock:
+                self.value += 1
+
     BATCH_SIZE = 500  # Dec 2021: delete_multi has a max size of 500
     NUM_THREADS = 3
     deletion_queue = Queue()
@@ -281,7 +291,7 @@ def _delete_all_values(client, key):
         while True:
             client.delete_multi(deletion_queue.get())
             shared_counter.increment()
-            LOGGER.info(
+            LOGGER.debug(
                 f"batch deletions completed: {shared_counter.value} ({shared_counter.value * BATCH_SIZE} total entries) & outstanding queue size: {deletion_queue.qsize()}"
             )
             deletion_queue.task_done()
