--- sdk/python/feast/infra/utils/generic_utils.py
@@ -1,11 +1,0 @@
-from threading import Lock
-
-
-class AtomicCounter(object):
-    def __init__(self):
-        self.value = 0
-        self.lock = Lock()
-
-    def increment(self):
-        with self.lock:
-            self.value += 1
